requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
yfinance
numpy
pandas
textblob
